Renamed MicrobeEditorKeyHandler and adapted CMakeList.txt
